do not show restore sample database button to non-admins (#22506)
* do not show restore sample database button to non-admins

* fix db list update after restoring the sample db
