+++ frontend/src/metabase/admin/databases/components/DatabaseList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DatabaseList";
